requirements.txt
PILLow
httplib2
oauth2client
qrcode
schedule
flask
flask_sqlalchemy